+++ addOns/ascanrulesAlpha/src/main/java/org/zaproxy/zap/extension/ascanrulesAlpha/TestUserAgent.java
@@ -21,18 +21,24 @@ package org.zaproxy.zap.extension.ascanrulesAlpha;
 
 import java.io.IOException;
 import java.net.UnknownHostException;
+import java.util.ArrayList;
+import java.util.List;
 import org.apache.commons.httpclient.URIException;
 import org.apache.log4j.Logger;
 import org.parosproxy.paros.Constant;
+import org.parosproxy.paros.control.Control;
 import org.parosproxy.paros.core.scanner.AbstractAppPlugin;
 import org.parosproxy.paros.core.scanner.Alert;
 import org.parosproxy.paros.core.scanner.Category;
 import org.parosproxy.paros.network.HttpHeader;
 import org.parosproxy.paros.network.HttpMessage;
 import org.parosproxy.paros.network.HttpRequestHeader;
+import org.zaproxy.zap.extension.custompayloads.CustomPayloadModel;
+import org.zaproxy.zap.extension.custompayloads.ExtensionCustomPayloads;
+import org.zaproxy.zap.extension.custompayloads.PluginWithConfigurablePayload;
 
 /** @author kniepdennis@gmail.com */
-public class TestUserAgent extends AbstractAppPlugin {
+public class TestUserAgent extends AbstractAppPlugin implements PluginWithConfigurablePayload {
 
     private static final Logger log = Logger.getLogger(TestUserAgent.class);
 
@@ -64,6 +70,7 @@ public class TestUserAgent extends AbstractAppPlugin {
         YAHOO_SLURP,
         I_PHONE_3
     };
+    public static final String USER_AGENT_PAYLOAD_CATEGORY = "User-Agent";
 
     private int originalResponseBodyHash;
 
@@ -75,6 +82,11 @@ public class TestUserAgent extends AbstractAppPlugin {
     @Override
     public String getName() {
         return Constant.messages.getString(MESSAGE_PREFIX + "name");
+    }
+
+    @Override
+    public String[] getDependency() {
+        return null;
     }
 
     @Override
@@ -98,15 +110,31 @@ public class TestUserAgent extends AbstractAppPlugin {
     }
 
     @Override
+    public void init() {}
+
+    @Override
     public void scan() {
         originalResponseBodyHash = getBaseMsg().getResponseBody().hashCode();
 
-        for (String userAgent : USER_AGENTS) {
+        for (CustomPayloadModel userAgentPayload : getUserAgentPayloads()) {
             if (isStop()) {
                 return;
             }
-            attack(userAgent);
+            attack(userAgentPayload.getPayload());
         }
+    }
+
+    private List<CustomPayloadModel> getUserAgentPayloads() {
+        ExtensionCustomPayloads extension =
+                Control.getSingleton()
+                        .getExtensionLoader()
+                        .getExtension(ExtensionCustomPayloads.class);
+
+        if (extension != null) {
+            return extension.getPayloadsByCategory(USER_AGENT_PAYLOAD_CATEGORY);
+        }
+
+        return getDefaultPayloads();
     }
 
     private void attack(String userAgent) {
@@ -155,5 +183,14 @@ public class TestUserAgent extends AbstractAppPlugin {
                 userAgent,
                 "",
                 newMsg);
+    }
+
+    @Override
+    public List<CustomPayloadModel> getDefaultPayloads() {
+        List<CustomPayloadModel> payloads = new ArrayList<>();
+        for (String userAgent : USER_AGENTS) {
+            payloads.add(new CustomPayloadModel(USER_AGENT_PAYLOAD_CATEGORY, userAgent));
+        }
+        return payloads;
     }
 }